Move the code from new menu to menu.js, simplify the code as I was following the wrong pattern before
