+++ src/components/menu/menu.scss
@@ -40,8 +40,8 @@ $dark-blue: #294374;
       width: 100%;
       padding: 0.75rem;
 
-      &:hover {
-        color: $dark-blue;
+      &:active {
+        color: #edeeef;
       }
     }
   }
@@ -249,7 +249,7 @@ $dark-blue: #294374;
         top: 2.5rem;
         left: 0;
         box-shadow: 0px 10px 21px -15px rgba(0, 0, 0, 0.2);
-
+        width: 100%;
         > ul {
           > li {
             padding: 0.5rem 0;
@@ -285,308 +285,3 @@ $dark-blue: #294374;
     }
   }
 }
-
-/* .menu {
-    clear: both;
-    overflow-y: auto;
-    background: $dark-gray;
-    color: #ffffff;
-    display: flex;
-    
-    margin: 0;
-    padding: 0.75rem;
-    list-style: none;
-    width: 100%;
-    position: fixed;
-    height: 100%;
-    top: 0;
-    left: 0;
-    display: none;
-
-    &.mobile-active {
-      display: block;
-    }
-
-    li {
-      margin-bottom: 0;
-      position: relative;
-
-      ul {
-        list-style: none;
-        padding: 0;
-        margin: 0;
-      }
-
-      &.has-subpages {
-        &:before {
-          content: '';
-          display: block;
-          width: 8px;
-          height: 24px;
-          background-image: url('/images/right-arrow.svg');
-          background-repeat: no-repeat;
-          background-size: contain;
-          background-position: right center;
-          position: absolute;
-          right: 0;
-          z-index: 10;
-          top: 11px;
-        }
-        &:after {
-          content: '';
-          position: absolute;
-          width: 48px;
-          height: 48px;
-          top: 0;
-          right: 0;
-          z-index: 11;
-        }
-        ul {
-          display: none;
-        }
-        &.active-submenu {
-          &:before {
-            transform: rotate(90deg) translate(0px, 5px);
-          }
-          > a {
-            font-weight: bold;
-          }
-          > .MainMenu__submenu-container > ul,
-          > ul {
-            display: block;
-            padding-left: 0.5rem;
-            margin-bottom: 1rem;
-          }
-        }
-      }
-    }
-
-    > li {
-      width: 100%;
-      box-sizing: border-box;
-
-      a, a:visited {
-        color: $color-white;
-        font-family: 'Alegreya Sans';
-        font-style: normal;
-        font-weight: 500;
-        font-size: 1rem;
-        text-decoration: none;
-        display: block;
-        width: 100%;
-        padding: 0.75rem 0;
-        border-bottom: 1px solid #f2f2f2;
-
-        &.active-link {
-          font-weight: bold;
-        }
-      }
-
-      &.has-subpages {
-        &.active-submenu {
-          top: 0;
-          left: 0;
-          margin: 0;
-          z-index: 20;
-        }
-        ul {
-          display: none;
-        }
-      }
-
-      .MainMenu__submenu-container {
-        width: 100%;
-        > ul {
-          @include layout-container-mixin;
-        }
-
-        display: block;
-      }
-
-      &:hover {
-        > .MainMenu__submenu-container {
-          display: block;
-        }
-      }
-    }
-  }
-
-  .menu-btn {
-    display: none;
-  }
-
-  .nav-icon:hover {
-    cursor: pointer;
-  }
-}
-
-@include media-breakpoint-up(md) {
-  .MainMenu {
-    width: 100%;
-    height: auto;
-    margin: 0;
-    background-color: $dark-blue;
-
-    &__mobile-open-button {
-      display: none;
-    }
-
-    &__mobile-close-button {
-      display: none;
-    }
-
-    .menu-icon {
-      display: none;
-    }
-
-    .menu {
-      @include layout-container-mixin;
-      background: transparent;
-      overflow-y: visible;
-      margin-bottom: 0;
-      box-shadow: 0px 10px 21px -15px rgba(0, 0, 0, 0.2);
-      position: relative;
-      top: auto;
-      left: auto;
-      padding: 0;
-      display: flex;
-      justify-content: flex-start;
-
-      li {
-        ul {
-          display: block;
-        }
-
-        &.has-subpages {
-          &:before,
-          &:after {
-            display: none;
-          }
-          &.active-submenu {
-            position: relative;
-            top: auto;
-            left: auto;
-            width: auto;
-            height: auto;
-
-            ul {
-              display: block;
-              padding: 0;
-            }
-          }
-          ul {
-            display: block;
-            padding: 0;
-          }
-        }
-      }
-
-      > li {
-        a, a:visited {
-          font-size: 1rem;
-          margin: 0;
-          display: flex;
-          height: auto;
-          align-items: center;
-          padding: 0.25rem 0;
-          border-bottom: none;
-        }
-
-        > a, > a:visited {
-          padding: 0.75rem;
-        }
-
-        width: auto;
-        display: inline-block;
-        padding: 0;
-        margin: 0;
-        &:hover {
-          background: $color-white;
-          > a {
-            color: $dark-gray;
-          }
-        }
-
-        .MainMenu__submenu-container {
-          background: $color-white;
-          position: fixed;
-          top: 2.5rem;
-          left: 0;
-          box-shadow: 0px 10px 21px -15px rgba(0, 0, 0, 0.2);
-          display: none;
-
-          > ul {
-            > li {
-              padding: 0.5rem 0;
-              display: flex;
-              flex-direction: column;
-              align-items: flex-start;
-              align-content: flex-start;
-              > a {
-                font-weight: bold;
-                font-size: 1.25rem;
-              }
-            }
-
-            a, a:visited {
-              color: $dark-gray;
-            }
-
-            > li > ul {
-              > li {
-                a, a:visited {
-                  height: auto;
-
-                  &:hover {
-                    text-decoration: underline;
-                  }
-                }
-              }
-            }
-          }
-        }
-
-        &:hover, .is-expanded {
-          > .MainMenu__submenu-container {
-            > ul {
-              padding: 1rem;
-              column-count: 3;
-
-              > li {
-                width: 100%;
-                display: inline-block;
-
-                > a {
-                  border-bottom: 1px solid $dark-gray;
-                  margin-bottom: 0.5rem;
-                  padding: 0.25rem 0;
-                }
-
-                > ul > li > ul > li {
-                  padding-left: 0.75rem;
-                  position: relative;
-                  margin-bottom: 0.15rem;
-
-                  &:before {
-                    color: $dark-gray;
-                    display: flex;
-                    content: '-';
-                    font-family: 'Arial';
-                    position: absolute;
-                    left: 0;
-                    top: 0;
-                    height: 100%;
-                    width: 0.75rem;
-                    align-items: flex-start;
-                    padding-top: 0.2rem;
-                    justify-content: flex-start;
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-} */
